BSP-116: Interfaces removed, commands updated, deserializer added.
