+++ .eslintrc.js
@@ -6,5 +6,8 @@ module.exports = {
         "eslint:recommended",
         "plugin:@typescript-eslint/recommended",
         "prettier"
-    ]
+    ],
+    env: {
+        jest: true
+    }
 };
