+++ src/commands/session/createSessionCommand.ts
@@ -1,10 +1,12 @@
-import { CreateSessionInterface } from "../../interfaces/session";
+interface CreateSessionInterface {
+    sessionToken: string;
+}
 
-export default class CreateSessionCommand {
+export default class CreateSessionCommand implements CreateSessionInterface {
     readonly sessionToken: string;
 
-    constructor(payload: CreateSessionInterface) {
-        const { sessionToken } = payload;
+    constructor(payload: unknown) {
+        const { sessionToken } = payload as CreateSessionInterface;
 
         if (typeof sessionToken !== "string") {
             throw new Error("Invalid session token");
