+++ src/commands/token/refreshTokenCommand.ts
@@ -1,13 +1,19 @@
-import { RefreshTokenInterface } from "../../interfaces/token";
+interface RefreshTokenInterface {
+    grant_type: string;
+    refresh_token: string;
+    basic_auth: string;
+    scope?: string | undefined;
+}
 
 export default class RefreshTokenCommand implements RefreshTokenInterface {
     readonly grant_type: string;
     readonly refresh_token: string;
     readonly basic_auth: string;
-    readonly scope: string | undefined;
+    readonly scope?: string | undefined;
 
-    constructor(payload: RefreshTokenInterface) {
-        const { grant_type, refresh_token, basic_auth, scope } = payload;
+    constructor(data: unknown) {
+        const { grant_type, refresh_token, basic_auth, scope } =
+            data as RefreshTokenInterface;
 
         if (typeof grant_type !== "string") {
             throw new Error("Invalid grant_type");
