+++ src/commands/userinfo/userinfoCommand.ts
@@ -1,10 +1,12 @@
-import { UserinfoInterface } from "../../interfaces/userinfo";
+interface UserinfoInterface {
+    bearerToken: string;
+}
 
 export default class UserinfoCommand implements UserinfoInterface {
     readonly bearerToken: string;
 
-    constructor(payload: UserinfoInterface) {
-        const { bearerToken } = payload;
+    constructor(payload: unknown) {
+        const { bearerToken } = payload as UserinfoInterface;
 
         if (typeof bearerToken !== "string") {
             throw new Error("Invalid bearerToken");
