+++ src/private/authnPrivateHandler.ts
@@ -1,11 +1,17 @@
-import { jsonCodec, PrivateNatsHandler } from "../nats/nats";
 import { headers, Subscription } from "nats";
+import { deserialize, jsonCodec, PrivateNatsHandler } from "../nats/nats";
 import { HANDLERS_SUBJECTS } from "../config";
-import { AuthnInterface, AuthnResponseInterface } from "../interfaces/authn";
 import AuthnCommand from "../commands/authn/authnCommand";
 import CreateSessionCommand from "../commands/session/createSessionCommand";
-import { CreateSessionResponse } from "../interfaces/session";
 import { authnService, sessionService } from "../di.config";
+import { CreateSessionResponse } from "./sessionPrivateHandler";
+import { ConnectorResponse } from "../network/config/connector";
+
+interface AuthnResponseInterface extends ConnectorResponse {
+    data: {
+        sessionToken: string;
+    };
+}
 
 export const authnPrivateHandlers: PrivateNatsHandler[] = [
     [
@@ -13,9 +19,10 @@ export const authnPrivateHandlers: PrivateNatsHandler[] = [
         async (subscription: Subscription): Promise<void> => {
             for await (const message of subscription) {
                 try {
-                    const data = jsonCodec.decode(
-                        message.data
-                    ) as AuthnInterface;
+                    const data = deserialize<AuthnCommand>(
+                        message.data,
+                        AuthnCommand
+                    );
                     const responseHeaders = headers();
 
                     const authnCommand = new AuthnCommand(data);
