+++ src/private/tokenPrivateHandler.ts
@@ -1,14 +1,36 @@
-import { jsonCodec, PrivateNatsHandler } from "../nats/nats";
 import { Subscription } from "nats";
+import { deserialize, jsonCodec, PrivateNatsHandler } from "../nats/nats";
 import { HANDLERS_SUBJECTS } from "../config";
 import RefreshTokenCommand from "../commands/token/refreshTokenCommand";
-import {
-    ExchangeTokenResponse,
-    RefreshTokenResponse,
-    TokenInterface
-} from "../interfaces/token";
 import ExchangeTokenCommand from "../commands/token/exchangeTokenCommand";
 import { tokenService } from "../di.config";
+import { ConnectorResponse } from "../network/config/connector";
+
+interface TokenInterface {
+    grant_type: string;
+    redirect_uri: string;
+    client_id: string;
+    client_secret: string;
+    code: string;
+    scope: string;
+    refresh_token: string;
+    basic_auth: string;
+}
+
+interface ExchangeTokenResponse extends ConnectorResponse {
+    data: {
+        access_token: string;
+        refresh_token: string;
+        scope: string;
+    };
+}
+
+interface RefreshTokenResponse extends ConnectorResponse {
+    data: {
+        access_token: string;
+        refresh_token?: string;
+    };
+}
 
 export const tokenPrivateHandlers: PrivateNatsHandler[] = [
     [
@@ -17,6 +39,7 @@ export const tokenPrivateHandlers: PrivateNatsHandler[] = [
             for await (const message of subscription) {
                 const data = jsonCodec.decode(message.data) as TokenInterface;
                 let refreshTokenCommand, exchangeCodeCommand;
+                let refreshTokenData, exchangeTokenData;
                 let tokenResponse;
 
                 try {
@@ -26,14 +49,25 @@ export const tokenPrivateHandlers: PrivateNatsHandler[] = [
 
                     switch (data.grant_type) {
                         case "refresh_token":
-                            refreshTokenCommand = new RefreshTokenCommand(data);
+                            refreshTokenData = deserialize<RefreshTokenCommand>(
+                                message.data,
+                                RefreshTokenCommand
+                            );
+                            refreshTokenCommand = new RefreshTokenCommand(
+                                refreshTokenData
+                            );
                             tokenResponse = (await tokenService.refreshToken(
                                 refreshTokenCommand
                             )) as RefreshTokenResponse;
                             break;
                         case "authorization_code":
+                            exchangeTokenData =
+                                deserialize<ExchangeTokenCommand>(
+                                    message.data,
+                                    ExchangeTokenCommand
+                                );
                             exchangeCodeCommand = new ExchangeTokenCommand(
-                                data
+                                exchangeTokenData
                             );
                             tokenResponse =
                                 (await tokenService.exchangeAuthorizationCode(
