+++ src/private/userinfoPrivateHandler.ts
@@ -1,9 +1,13 @@
-import { jsonCodec, PrivateNatsHandler } from "../nats/nats";
 import { Subscription } from "nats";
+import { jsonCodec, PrivateNatsHandler } from "../nats/nats";
 import { HANDLERS_SUBJECTS } from "../config";
 import UserinfoCommand from "../commands/userinfo/userinfoCommand";
-import { UserinfoResponse } from "../interfaces/userinfo";
 import { userinfoService } from "../di.config";
+import { ConnectorResponse } from "../network/config/connector";
+
+interface UserinfoResponse extends ConnectorResponse {
+    data: Record<never, unknown>;
+}
 
 export const userinfoPrivateHandlers: PrivateNatsHandler[] = [
     [
