+++ src/public/authorizePublicHandler.ts
@@ -1,10 +1,10 @@
+import { headers, Subscription } from "nats";
 import {
     AirlockPayload,
     getConnection,
     jsonCodec,
     PublicNatsHandler
 } from "../nats/nats";
-import { Subscription } from "nats";
 import { HANDLERS_SUBJECTS, SERVICE_NAME } from "../config";
 
 export const authorizePublicHandlers: PublicNatsHandler[] = [
@@ -15,15 +15,17 @@ export const authorizePublicHandlers: PublicNatsHandler[] = [
             for await (const message of subsciption) {
                 try {
                     const natsConnection = getConnection();
-
                     const { query } = jsonCodec.decode(
                         message.data
                     ) as AirlockPayload;
 
                     const response = await natsConnection.request(
                         `${SERVICE_NAME}.${HANDLERS_SUBJECTS.AUTHORIZE}`,
-                        jsonCodec.encode(query),
-                        { timeout: 6000, headers: message.headers }
+                        jsonCodec.encode({
+                            ...(query as Record<never, unknown>),
+                            cookie: headers().get("cookie")
+                        }),
+                        { timeout: 6000 }
                     );
 
                     message.respond(response.data);
