+++ src/services/oktaAuthorizeService.ts
@@ -2,14 +2,23 @@ import ConnectorInterface, {
     ConnectorResponse
 } from "../network/config/connector";
 import { AUTHORIZE_ENDPOINT, INTERNAL_REDIRECT_URI } from "../config";
-import { AuthorizeServiceInterface } from "../interfaces/authorize";
 import AuthorizeCommand from "../commands/authorize/authorizeCommand";
 
+interface AuthorizeServiceInterface {
+    authorize: (
+        authorizeCommand: AuthorizeCommand
+    ) => Promise<ConnectorResponse>;
+}
+
 export default class OktaAuthorizeService implements AuthorizeServiceInterface {
-    private readonly _restConnector: ConnectorInterface;
+    private readonly restConnector: ConnectorInterface;
 
     constructor(restConnector: ConnectorInterface) {
-        this._restConnector = restConnector;
+        if (!restConnector) {
+            throw new Error("No valid REST connector.");
+        }
+
+        this.restConnector = restConnector;
     }
 
     public async authorize(
@@ -17,24 +26,21 @@ export default class OktaAuthorizeService implements AuthorizeServiceInterface {
     ): Promise<ConnectorResponse> {
         const { state, client_id, scope, cookie, sessionToken } =
             authorizeCommand;
+        const parameters = {
+            client_id: client_id,
+            redirect_uri: INTERNAL_REDIRECT_URI,
+            response_type: "code",
+            scope: scope,
+            prompt: "none",
+            state: state,
+            nonce: "RandomNonce",
+            ...(sessionToken && { sessionToken: sessionToken })
+        };
 
-        return this._restConnector.get(
-            AUTHORIZE_ENDPOINT,
-            {
-                client_id: client_id,
-                redirect_uri: INTERNAL_REDIRECT_URI,
-                response_type: "code",
-                scope: scope,
-                prompt: "none",
-                state: state,
-                nonce: "RandomNonce",
-                ...(sessionToken && { sessionToken: sessionToken })
-            },
-            {
-                headers: {
-                    ...(cookie && { cookie: cookie })
-                }
+        return this.restConnector.get(AUTHORIZE_ENDPOINT, parameters, {
+            headers: {
+                ...(cookie && { cookie: cookie })
             }
-        );
+        });
     }
 }
