+++ src/services/oktaUserinfoService.ts
@@ -2,14 +2,21 @@ import ConnectorInterface, {
     ConnectorResponse
 } from "../network/config/connector";
 import { USERINFO_ENDPOINT } from "../config";
-import { UserinfoServiceInterface } from "../interfaces/userinfo";
 import UserinfoCommand from "../commands/userinfo/userinfoCommand";
 
+interface UserinfoServiceInterface {
+    userinfo: (userinfoCommand: UserinfoCommand) => Promise<ConnectorResponse>;
+}
+
 export default class OktaUserinfoService implements UserinfoServiceInterface {
-    private readonly _restConnector: ConnectorInterface;
+    private readonly restConnector: ConnectorInterface;
 
     constructor(restConnector: ConnectorInterface) {
-        this._restConnector = restConnector;
+        if (!restConnector) {
+            throw new Error("No valid REST connector.");
+        }
+
+        this.restConnector = restConnector;
     }
 
     public async userinfo(
@@ -17,7 +24,7 @@ export default class OktaUserinfoService implements UserinfoServiceInterface {
     ): Promise<ConnectorResponse> {
         const { bearerToken } = userinfoCommand;
 
-        return this._restConnector.get(USERINFO_ENDPOINT, undefined, {
+        return this.restConnector.get(USERINFO_ENDPOINT, undefined, {
             headers: {
                 authorization: bearerToken
             }
